requirements.txt
Flask==2.3.2
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
openai==1.6.1
matplotlib>=3.8.0